reorganized js and css for admin side using new webpack options
