--- inc/custom-blocks/core-extensions/header/header-block.css
@@ -1,104 +1,0 @@
-.heading-reset.has-small-font-size h1 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h1 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h1 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h1 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h1 {
-  font-size: 48px !important;
-}
-
-.heading-reset.has-small-font-size h2 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h2 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h2 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h2 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h2 {
-  font-size: 48px !important;
-}
-
-.heading-reset.has-small-font-size h3 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h3 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h3 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h3 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h3 {
-  font-size: 48px !important;
-}
-
-.heading-reset.has-small-font-size h4 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h4 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h4 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h4 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h4 {
-  font-size: 48px !important;
-}
-
-.heading-reset.has-small-font-size h5 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h5 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h5 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h5 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h5 {
-  font-size: 48px !important;
-}
-
-.heading-reset.has-small-font-size h6 {
-  font-size: 13px !important;
-}
-.heading-reset.has-normal-font-size h6 {
-  font-size: 16px !important;
-}
-.heading-reset.has-medium-font-size h6 {
-  font-size: 20px !important;
-}
-.heading-reset.has-large-font-size h6 {
-  font-size: 36px !important;
-}
-.heading-reset.has-huge-font-size h6 {
-  font-size: 48px !important;
-}
-
-.heading-reset h1,
-.heading-reset h2,
-.heading-reset h3,
-.heading-reset h4,
-.heading-reset h5,
-.heading-reset h6 {
-  font-weight: normal;
-}
