+++ src/js/editor/slider-block.js
