+++ src/js/emma-admin.js
@@ -1,3 +1,6 @@
 /*--------------------------------------------------------------
 # Emma - Admin Scripts
 --------------------------------------------------------------*/
+import "./editor/header-block.js";
+import "./editor/responsive-grid-block.js";
+import "./editor/slider-block.js";
