+++ src/sass/editor/responsive-grid-block.css
@@ -1,3 +1,5 @@
+// @TODO: BEGIN Convert to SASS
+
 .wp-block-emma-responsive-grid > .block-editor-inner-blocks > .block-editor-block-list__layout {
   display: grid;
 }
