+++ src/sass/editor/slider-block.css
