+++ src/sass/emma-admin.scss
@@ -1,3 +1,44 @@
 /*--------------------------------------------------------------
 # Emma - Admin Styles
 --------------------------------------------------------------*/
+
+/*--------------------------------------------------------------
+# Bootstrap Functions (vendor/bootstrap/bootstrap-reboot.scss)
+--------------------------------------------------------------*/
+@import "vendor/bootstrap/_functions.scss";
+
+/*--------------------------------------------------------------
+# Variables
+--------------------------------------------------------------*/
+@import "variables/variables";
+@import "mixins/mixins";
+
+/*--------------------------------------------------------------
+# Bootstrap Reboot (vendor/bootstrap/bootstrap-reboot.scss)
+--------------------------------------------------------------*/
+@import "vendor/bootstrap/_variables.scss";
+@import "vendor/bootstrap/_mixins.scss";
+@import "vendor/bootstrap/_reboot.scss";
+
+/*--------------------------------------------------------------
+# Gutenberg
+--------------------------------------------------------------*/
+@import "gutenberg/gutenberg";
+
+/*--------------------------------------------------------------
+# Overrides and Custom Blocks
+--------------------------------------------------------------*/
+@import "editor/overrides";
+@import "editor/header-block";
+@import "editor/responsive-grid-block";
+@import "editor/slider-block";
+
+/*--------------------------------------------------------------
+# Modules
+--------------------------------------------------------------*/
+@import "modules/utilities";
+
+/*--------------------------------------------------------------
+# Forms
+--------------------------------------------------------------*/
+@import "forms/forms";
